Rename profile_model -> profile
