+++ archinstall/lib/models/__init__.py
@@ -36,7 +36,7 @@ from .locale import LocaleConfiguration
 from .mirrors import CustomRepository, MirrorConfiguration, MirrorRegion
 from .network_configuration import NetworkConfiguration, Nic, NicType
 from .packages import LocalPackage, PackageSearch, PackageSearchResult, Repository
-from .profile_model import ProfileConfiguration
+from .profile import ProfileConfiguration
 from .users import PasswordStrength, User
 
 __all__ = [
