+++ archinstall/lib/profile/profiles_handler.py
@@ -14,7 +14,7 @@ from archinstall.lib.translationhandler import tr
 
 from ...default_profiles.profile import GreeterType, Profile
 from ..hardware import GfxDriver
-from ..models.profile_model import ProfileConfiguration
+from ..models.profile import ProfileConfiguration
 from ..networking import fetch_data_from_url, list_interfaces
 from ..output import debug, error, info
