+++ examples/full_automated_installation.py
@@ -18,7 +18,7 @@ from archinstall.lib.models.device_model import (
 	Size,
 	Unit,
 )
-from archinstall.lib.models.profile_model import ProfileConfiguration
+from archinstall.lib.models.profile import ProfileConfiguration
 from archinstall.lib.models.users import Password, User
 from archinstall.lib.profile.profiles_handler import profile_handler
